api/requirements/local.txt
# Local development dependencies go here
-r base.txt
django-mockups  #Genera datos de prueba
django-extensions #Agrega mas funcionalidad al shell de django
ipdb
